reorganizing file structure and starting styling
