+++ client/src/App.css
@@ -1,0 +1,4 @@
+.main{
+    display: flex;
+    flex-flow: row;
+}
