--- client/src/components/Sidebar.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./Sidebar.css";
-
-const Sidebar = (props) => {
-    return (
-        <>
-            <button
-            type = "button"
-            onClick = {() => props.setShowInputForm(!props.showInputForm)} //toggle form showing
-            >New Date</button>
-        </>
-    );
-}
-export default Sidebar;
